refactor(settingsView): move state-setting write orchestration to controllers
applyStateSettingUpdate/postStateSettingSnapshot performed real orchestration
(workspace-guard branching, persistence I/O, an approval-policy side effect)
but lived in src/shared/settingsView/handlers/, where the rule is host-neutral
orchestration belongs in src/controllers/ instead. Split the file: the pure,
I/O-free catalog resolution (resolveStateSettingWrite) stays in src/shared/;
the write path moves to a new StateSettingUpdateController alongside its
sibling settingsView controllers.
