+++ src/shared/settingsView/handlers/stateSettingWrite.ts
@@ -1,26 +1,18 @@
 // Shared routing decision for the generic `UPDATE_STATE_SETTING` command.
 //
-// The extension and desktop hosts share persistence through settingsAccess and
-// own only the post-write side effects for each outbound snapshot. This
-// resolver owns the subtle boundary rules once:
+// This resolver owns the subtle catalog boundary rules once:
 //   - a value-less message is a no-op (the catalog schemas `.prefault()`, so
 //     parsing `undefined` would silently write a default),
 //   - null explicitly resets a setting while an omitted value remains a no-op,
 //   - only catalog rows tagged for a settings-view snapshot are writable.
+//
+// It performs no I/O and has no host dependency, so it stays in `src/shared`.
+// The write orchestration that consumes this decision (persist, guard, fire
+// the approval-policy side effect) is host-neutral but not pure, so it lives
+// in `@controllers/settingsView/StateSettingUpdateController` instead.
 
 import {
-  TEXRA_APPROVAL_POLICY_CONFIG_KEY,
-  readPersistedTexraApprovalPolicy,
-  type TexraApprovalPolicy,
-} from '@shared/approvalPolicy';
-import {
-  resetSetting,
-  writeSetting,
-  type SettingsStores,
-} from '@shared/config/settingsAccess';
-import {
   settingsViewSettingByKey,
-  type SettingsViewSnapshot,
   type SettingsViewStateSettingEntry,
 } from '@shared/schemas/stateSettings';
 
@@ -29,7 +21,7 @@ import {
  * when the message must be ignored (value-less, unknown key, or a catalog row
  * this settings view does not own).
  */
-type StateSettingWrite =
+export type StateSettingWrite =
   | {
       readonly kind: 'write';
       readonly entry: SettingsViewStateSettingEntry;
@@ -62,96 +54,4 @@ export function resolveStateSettingWrite(
   const parsed = entry.schema.safeParse(value);
   if (!parsed.success) return { kind: 'rejected', entry, error: parsed.error };
   return { kind: 'write', entry, value: parsed.data };
-}
-
-/** Outcome of {@link applyStateSettingUpdate}, for host-specific UI feedback. */
-export type StateSettingUpdateResult =
-  | { readonly kind: 'ignored' }
-  | {
-      readonly kind: 'rejected';
-      readonly entry: SettingsViewStateSettingEntry;
-      readonly error: Error;
-    }
-  | {
-      readonly kind: 'workspace-required';
-      readonly entry: SettingsViewStateSettingEntry;
-    }
-  | { readonly kind: 'applied'; readonly entry: SettingsViewStateSettingEntry }
-  | {
-      readonly kind: 'failed';
-      readonly entry: SettingsViewStateSettingEntry;
-      readonly error: unknown;
-    };
-
-export interface StateSettingUpdatePorts {
-  readonly stores: SettingsStores;
-  /**
-   * Extension-only guard: a workspace-target config write needs an open
-   * workspace folder. Hosts without that constraint (desktop, CLI) omit this.
-   */
-  readonly requiresOpenWorkspace?: (
-    entry: SettingsViewStateSettingEntry,
-  ) => boolean;
-  /** Applies the approval-policy side effect when that setting changes. */
-  readonly onApprovalPolicyChanged?: (policy: TexraApprovalPolicy) => void;
-}
-
-/**
- * Host-neutral write path for a generic `UPDATE_STATE_SETTING` message:
- * resolve, guard, persist, and apply the approval-policy side effect. Callers
- * own all UI feedback and the outbound snapshot rebroadcast — this performs
- * only the decision and the write.
- */
-export async function applyStateSettingUpdate(
-  key: string,
-  value: unknown,
-  ports: StateSettingUpdatePorts,
-): Promise<StateSettingUpdateResult> {
-  const write = resolveStateSettingWrite(key, value);
-  if (!write) return { kind: 'ignored' };
-  if (write.kind === 'rejected') {
-    return { kind: 'rejected', entry: write.entry, error: write.error };
-  }
-  if (
-    write.entry.store === 'config' &&
-    write.entry.configTarget !== 'global' &&
-    ports.requiresOpenWorkspace?.(write.entry)
-  ) {
-    return { kind: 'workspace-required', entry: write.entry };
-  }
-  try {
-    await (write.kind === 'reset'
-      ? resetSetting(write.entry, ports.stores)
-      : writeSetting(write.entry, write.value, ports.stores));
-    if (write.entry.key === TEXRA_APPROVAL_POLICY_CONFIG_KEY) {
-      ports.onApprovalPolicyChanged?.(
-        write.kind === 'reset'
-          ? readPersistedTexraApprovalPolicy((k, fallback) =>
-              ports.stores.config.get(k, fallback),
-            )
-          : (write.value as TexraApprovalPolicy),
-      );
-    }
-    return { kind: 'applied', entry: write.entry };
-  } catch (error) {
-    return { kind: 'failed', entry: write.entry, error };
-  }
-}
-
-/**
- * Rebroadcast posters for every {@link SettingsViewSnapshot}. A `Record` (not a
- * `switch`) so a new snapshot variant fails the object-literal check at both
- * call sites instead of silently falling through a `default`.
- */
-export type SettingsSnapshotPosters = Record<
-  SettingsViewSnapshot,
-  () => void | Promise<void>
->;
-
-/** Rebroadcasts the settings-view snapshot a state-setting write belongs to. */
-export async function postStateSettingSnapshot(
-  snapshot: SettingsViewSnapshot,
-  posters: SettingsSnapshotPosters,
-): Promise<void> {
-  await posters[snapshot]();
 }
